+++ CHANGELOG.md
@@ -336,3 +336,8 @@ available messages for post
 * `PostPublisher` channel `postPublishedChannel` now have list of messages as value for notification
 
 ### 1.2.7
+
+* `kotlin` version `1.3.21` -> `1.3.30`
+* `kotlin coroutines` version `1.1.1` -> `1.2.0`
+* `kotlin serialization` version `0.10.0` -> `0.11.0`
+* `TelegramBotAPI` version `0.12.5` -> `0.12.6`